+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Camera can be snapped to pixels for pixel art games.
 * Sprite Renderer can be snapped to pixels for pixel art games.
 * Can edit processor parameters 'Texture Format' and 'Premultiply Alpha' for font assets.
+* Text Renderer can be snapped to pixels for pixel art games.
 
 ## Refactors